Refactor Microsoft Graph client authentication
This commit removes the `DestinationClientCreator` class and introduces the `DestinationTenantGraphClient` class to manage the creation of `GraphServiceClient` instances and token acquisition. The `TokenProvider` class has been updated to use a more generic approach with `sourceClientID` and `authority`. The `DestinationTenantSettings` form is now updated to utilize the new graph client for authentication. Additional using directives have been added to support the new structure, enhancing code maintainability and clarity.